Rename near to neard
