+++ CODEOWNERS
@@ -8,4 +8,4 @@
 /genesis-tools/ @nearmax
 
 /core/ @nearmax @evgenykuzyakov @SkidanovAlex @bowenwang1996
-/near/ @nearmax @evgenykuzyakov @SkidanovAlex @bowenwang1996
+/neard/ @nearmax @evgenykuzyakov @SkidanovAlex @bowenwang1996
