+++ Cargo.toml
@@ -29,7 +29,7 @@ members = [
     "test-utils/testlib",
     "test-utils/loadtester",
     "test-utils/state-viewer",
-    "near/",
+    "neard/",
     "tools/rpctypegen/core",
     "tools/rpctypegen/macro",
     "genesis-tools/genesis-csv-to-json",
@@ -57,7 +57,7 @@ node-runtime = { path = "./runtime/runtime" }
 near-jsonrpc = { path = "./chain/jsonrpc" }
 near-network = { path = "./chain/network" }
 
-near = { path = "./near"}
+neard = { path = "./neard"}
 
 testlib = { path = "./test-utils/testlib" }
 
@@ -84,11 +84,11 @@ opt-level = 3 # BLS library is too slow to use in debug
 opt-level = 3 # bs58 library is too slow to use in debug
 
 [features]
-expensive_tests = ["near/expensive_tests"]
+expensive_tests = ["neard/expensive_tests"]
 regression_tests = []
 old_tests = []
-adversarial = ["near/adversarial", "near-jsonrpc/adversarial"]
-metric_recorder = ["near/metric_recorder"]
+adversarial = ["neard/adversarial", "near-jsonrpc/adversarial"]
+metric_recorder = ["neard/metric_recorder"]
 
 [patch.crates-io]
 actix-rt = { git = "https://github.com/actix/actix-net", rev="602db1779eb51d60e0fe5a33d725d1d7fdf540fd" }
