+++ chain/chain/src/chain.rs
@@ -64,7 +64,7 @@ const ACCEPTABLE_TIME_DIFFERENCE: i64 = 12 * 10;
 /// Over this block height delta in advance if we are not chunk producer - route tx to upcoming validators.
 pub const TX_ROUTING_HEIGHT_HORIZON: BlockHeightDelta = 4;
 
-/// Private constant for 1 NEAR (copy from near/config.rs) used for reporting.
+/// Private constant for 1 NEAR (copy from neard/config.rs) used for reporting.
 const NEAR_BASE: Balance = 1_000_000_000_000_000_000_000_000;
 
 /// Number of epochs for which we keep store data
