+++ genesis-tools/genesis-csv-to-json/Cargo.toml
@@ -10,7 +10,7 @@ csv = "1.1.1"
 serde = { version = "1.0", features = ["derive"] }
 clap = "2.33.0"
 
-near = { path = "../../near" }
+neard = { path = "../../neard" }
 near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
