+++ genesis-tools/genesis-populate/Cargo.toml
@@ -11,7 +11,7 @@ indicatif = "0.13.0"
 clap = "2.33.0"
 tempdir = "0.3"
 
-near = { path = "../../near" }
+neard = { path = "../../neard" }
 near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
