+++ genesis-tools/keypair-generator/src/main.rs
@@ -3,7 +3,7 @@ use std::path::{Path, PathBuf};
 
 use clap::{App, AppSettings, Arg, SubCommand};
 
-use near::get_default_home;
+use neard::get_default_home;
 use near_crypto::{InMemorySigner, KeyType, SecretKey, Signer};
 
 fn generate_key_to_file(account_id: &str, key: SecretKey, path: PathBuf) {
@@ -88,7 +88,7 @@ fn main() {
                 let account_id =
                     account_id.expect("Account id must be specified if --generate-config is used");
                 let mut path = home_dir.to_path_buf();
-                path.push(near::config::VALIDATOR_KEY_FILE);
+                path.push(neard::config::VALIDATOR_KEY_FILE);
                 generate_key_to_file(account_id, key, path);
             }
         }
@@ -97,7 +97,7 @@ fn main() {
             println!("PK: {}", key.public_key());
             if generate_config {
                 let mut path = home_dir.to_path_buf();
-                path.push(near::config::NODE_KEY_FILE);
+                path.push(neard::config::NODE_KEY_FILE);
                 generate_key_to_file("", key, path);
             }
         }
