+++ neard/res/testnet_genesis_config.json
