+++ neard/src/near.rs
@@ -1,0 +1,5 @@
+use neard::neard_main;
+
+fn main() {
+    neard_main();
+}
