+++ runtime/runtime-params-estimator/src/testbed.rs
@@ -1,5 +1,5 @@
 use borsh::BorshDeserialize;
-use near::get_store_path;
+use neard::get_store_path;
 use near_primitives::receipt::Receipt;
 use near_primitives::transaction::{ExecutionStatus, SignedTransaction};
 use near_primitives::types::{Gas, MerkleHash, StateRoot};
