+++ test-utils/state-viewer/Cargo.toml
@@ -17,5 +17,5 @@ near-store = { path = "../../core/store" }
 node-runtime = { path = "../../runtime/runtime" }
 near-chain = { path = "../../chain/chain" }
 near-network = { path = "../../chain/network" }
-near = { path = "../../near" }
+neard = { path = "../../neard" }
