+++ test-utils/testlib/src/lib.rs
@@ -3,7 +3,6 @@ use std::sync::Arc;
 use actix::Addr;
 use tempdir::TempDir;
 
-use near::{config::GenesisExt, load_test_config, start_with_config, NightshadeRuntime};
 use near_chain::{Chain, ChainGenesis, DoomslugThresholdMode};
 use near_chain_configs::Genesis;
 use near_client::{ClientActor, ViewClientActor};
@@ -13,6 +12,7 @@ use near_primitives::hash::CryptoHash;
 use near_primitives::test_utils::init_integration_logger;
 use near_primitives::types::{BlockHeight, BlockHeightDelta, NumSeats, NumShards, ShardId};
 use near_store::test_utils::create_test_store;
+use neard::{config::GenesisExt, load_test_config, start_with_config, NightshadeRuntime};
 
 pub mod actix_utils;
 pub mod fees_utils;
