+++ test-utils/testlib/src/node/thread_node.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use near::{start_with_config, NearConfig};
+use neard::{start_with_config, NearConfig};
 use near_chain_configs::Genesis;
 use near_crypto::{InMemorySigner, KeyType, Signer};
 use near_primitives::types::AccountId;
