+++ tools/restaked/Cargo.toml
@@ -12,6 +12,6 @@ env_logger = "0.7"
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 near-jsonrpc-client = { path = "../../chain/jsonrpc/client" }
-near = { path = "../../near" }
+neard = { path = "../../neard" }
 
 testlib = { path = "../../test-utils/testlib" }
